adaptix/services/manufacturing/requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-jose[cryptography]
requests
django-cors-headers
dj-database-url
PyJWT>=2.8.0
celery
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
pytest==8.0.0
pytest-django==4.8.0
pytest-mock==3.12.0
drf-spectacular
whitenoise
django-filter>=23.2
